backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async Task Queue
celery[redis]==5.3.4
redis==5.0.1

# AI & ML
google-generativeai==0.3.2

# Video Processing
opencv-python==4.9.0.80
moviepy==1.0.3

# Vector Database (for RAG)
chromadb==0.4.22

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1

# Utilities
requests==2.31.0

# Google Drive Integration
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Groq STT Integration
groq==0.4.2
